Use dispatch instead of eventbus for supervisor events (#46986)
